refactor(conversation): extract ConversationStepRunner — Wave R complete
R3 step 2, completing Wave R. All three monoliths are decomposed:

  GroupConversationService  4,417 -> 1,422
  AgentOrchestrator         2,725 -> 1,163
  ConversationService       2,735 -> 1,174

Checkstyle is down from 8 violations to 6, and every remaining one is a
pre-existing LineLength in a file this branch never touched. No FileLength
violation is left anywhere in the codebase.

With the HITL cluster gone, what remained in ConversationService was two
things wearing one name: the public IConversationService surface (start, say,
read, undo/redo, access checks) and the machinery that executes a turn. This
class is the machinery; the facade is the surface.

The processing gauge and its release token are the delicate part.
ProcessingTurn increments processingConversationCount on construction and
releases exactly once, and every path out of a turn -- completion, timeout,
abandonment, a pre-submission throw -- must release it or the gauge drifts
upward forever and the graceful-shutdown drain waits on finished turns. Both
the token type and releaseTurn stay on the facade, shared by reference,
because say/sayStreaming create the token before the runner sees it.

Two bounds errors were caught before they could hide: the first attempt
sliced into getAgent (no javadoc, so the walk-back heuristic overshot), the
second overshot by two lines. Both produced immediate parse errors rather
than silently valid-but-wrong code, which is the argument for cutting exact
ranges and letting the compiler check the seam.

2,640 tests green.
